+++ test/spec/ci_spec.rb
@@ -7,4 +7,12 @@ describe "A PHP application on Heroku CI" do
 			expect(test_run.output).to match("mockery/mockery")
 		end
 	end
+	
+	it "has zend.assertions enabled" do
+		app = new_ci_app_with_stack_and_platrepo('test/fixtures/ci/zendassert', allow_failure: true)
+		app.run_ci do |test_run|
+			expect(test_run.status).to eq :failed
+			expect(test_run.output).to match("AssertionError")
+		end
+	end
 end